requirements.txt
fastapi
uvicorn[standard]
beanie
motor
pydantic-settings
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
email-validator
